remove webpack and change scripts, add circular plugin to production
