+++ frontend/babel.config.json
@@ -4,6 +4,5 @@
     "@babel/plugin-transform-runtime",
     "@babel/plugin-transform-modules-commonjs"
   ],
-  "only": ["*.{mjs,js}"],
-  "code": false
+  "only": ["*.{mjs,js}"]
 }
