+++ frontend/esbuild.config.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line
 const esbuild = require("esbuild");
 const path = require("path");
-const madge = require("madge");
+const circular = require("./esbuild-circular-plugin");
 
 /**
  * @type {esbuild.BuildOptions}
@@ -23,38 +23,7 @@ const buildOptions = {
     global: "window",
   },
   logLevel: "warning",
-  plugins: [
-    {
-      name: "circulars",
-      setup(build) {
-        build.onStart(async () => {
-          const madgeInstance = await madge(
-            build.initialOptions.entryPoints[0],
-            {
-              includeNpm: true,
-              tsConfig: "tsconfig.json",
-            }
-          );
-
-          const circulars = madgeInstance.circular();
-
-          return {
-            errors: circulars.map((v) => ({
-              text: "Circular Dependency Found",
-              location: {
-                file: v[0] ?? "",
-              },
-            })),
-            warnings: [
-              {
-                text: `Found ${circulars.length} circular dependencies`,
-              },
-            ],
-          };
-        });
-      },
-    },
-  ],
+  plugins: [circular],
 };
 
 module.exports = buildOptions;
