+++ frontend/package.json
@@ -14,13 +14,9 @@
     "dep-graph": "madge -c -i \"dep-graph.png\" ./src/scripts",
     "build": "npx gulp build",
     "build:live": "npx gulp build-production",
-    "build:live:esbuild": "npx gulp build-production-esbuild",
-    "build:webpack": "webpack --config webpack.config.js",
     "start:dev": "npm run build && concurrently --kill-others \"npx gulp watch\" \"firebase serve --only hosting\"",
     "deploy:live": "npm run build:live && firebase deploy -P live --only hosting",
-    "tsc": "tsc",
-    "build:esbuild": "npx gulp build-esbuild",
-    "start:dev:esbuild": "npm run build:esbuild && concurrently --kill-others \"npx gulp watch-esbuild\" \"firebase serve --only hosting\""
+    "tsc": "tsc"
   },
   "engines": {
     "npm": "8.1.2"
@@ -41,7 +37,6 @@
     "del": "6.0.0",
     "esbuild": "0.14.23",
     "esbuild-plugin-babel": "github:schloerke/esbuild-plugin-babel#patch-2",
-    "eslint-webpack-plugin": "3.1.1",
     "gulp": "4.0.2",
     "gulp-concat": "2.6.1",
     "gulp-eslint-new": "1.3.0",
@@ -53,9 +48,7 @@
     "ts-loader": "9.2.6",
     "typescript": "4.5.5",
     "vinyl-buffer": "1.0.1",
-    "vinyl-paths": "3.0.1",
-    "webpack": "5.68.0",
-    "webpack-cli": "4.9.2"
+    "vinyl-paths": "3.0.1"
   },
   "dependencies": {
     "axios": "0.21.2",
